+++ .template-lintrc.js
@@ -1,12 +1,17 @@
 /* eslint-env node */
+
 'use strict';
 
 module.exports = {
-  extends: 'recommended',
+    extends: 'recommended',
 
-  rules: {
-      'block-indentation': 4,
-      'bare-strings': true,
-      'nested-interactive': false,
-  }
+    rules: {
+        'block-indentation': 4,
+        'bare-strings': true,
+        'nested-interactive': false,
+    },
+
+    ignore: [
+        '**/lib/handbook/**/*',
+    ],
 };