+++ app/router.ts
@@ -50,6 +50,7 @@ const Router = EmberRouter.extend({
 const {
     engines: {
         collections,
+        handbook,
     },
 } = config;
 
@@ -66,6 +67,9 @@ Router.map(function() {
 
     if (collections.enabled) {
         this.mount('collections');
+    }
+    if (handbook.enabled) {
+        this.mount('handbook');
     }
 
     /*
@@ -94,8 +98,8 @@ Router.map(function() {
     this.route('resolve-guid', { path: '/:guid' });
 
     // Error routes
-    this.route('not-found', { path: '*path' });
     this.route('error-no-api', { path: '*no_api_path' });
+    this.route('not-found', { path: '*path' });
 });
 
 /* eslint-enable array-callback-return */