+++ config/environment.d.ts
@@ -139,6 +139,14 @@ declare const config: {
         collections: {
             enabled: boolean;
         };
+        handbook: {
+            enabled: boolean;
+            docGenerationEnabled: boolean;
+        };
+    };
+    lintOnBuild: boolean;
+    'ember-cli-tailwind'?: {
+        shouldIncludeStyleguide: boolean,
     };
 };